Move Drivetrain to ftc module
